[feat] Integrate AI Insights platform tabs for YouTube, Instagram, and Gmail
