+++ src/app/dashboard/_components/content-analytics/platforms/index.ts
@@ -1,0 +1,3 @@
+export { YouTubePlatform } from './YouTubePlatform';
+export { InstagramPlatform } from './InstagramPlatform';
+export { GmailPlatform } from './GmailPlatform';
