+++ src/app/dashboard/ai-insights/_components/hooks/index.ts
@@ -1,0 +1,3 @@
+export { useYouTubeInsights } from './useYouTubeInsights';
+export { useInstagramInsights } from './useInstagramInsights';
+export { useGmailInsights } from './useGmailInsights';
